Set up Elements, Node files
I set up the files for the Element and Node classes. These files don't have anything in them right now, this commit is just to establish the file structure of the project.

I also included a Forward_Declarations.h file. This file, as the name suggests, has forward declarations for the Node and Element classes. I doubt that I'll actually need this, it's moslty just a back up (just in case).