Added array class, Wrote Node class, made lots of changes
I began writing the node class, but this quickly caused issues. I realized that I was using a lottttt of 3 element arrays (of various types). To make this a lot easier, I wrote an "Array_3" class. This is simply a template class for 3 element arrays.

For the Array_3 class, I defined access methods (both () and []... () is safter but has an extra check, [] doesn't check but is faster). And a few constructor.

I also wrote basic code for the node class. To keep as much as possible private, I used member intialization lists (see node.h). I also defined a few getter methods (for the current and original position) + a setter (to update the position).

This code compiles but has not been tested.