Wrote tests
I wrote test functions. These functions basically just test that the Node class is working as intdned (whcih it appears to be).

I also modified the print method of the Node class so that it only prints the BC's if there are any.

Finally, I added a constructor for Array_3 that accepts 3 inputs. this allows me to pass collections of basic variables as Array_3 objects (see Tests.cc:40).