Added makefile + Added Array source file.
I redid the structure of the code. Rather than including everything in Main.cc, I am now using a make file. As a result, I had to properly set up each Source File's dependancies. This accounts for the bulk of the changes made in this commit.

I also found a way to move the declarations of the Array class methods into a source file. Since Array_3 is a template class, the header file needs to store all relevant information about the class (otherwise it won't be able to use the template to generate classes). To make this work, I moved all function delcarations into a Source file (Array.cc) and then included this source file at the end of the Array header file (That way, technically evertyhing is done in the headerfile, while still allowing me to write a separate source file).

The Makefile includes two variables (for the compiler and the flags). The code compiles and runs!

In my next commit, I plan on implementing the Element class.