Implemented Elements class (I think)
I wrote and implemented the Elements class. I basically followed the implementation plan, but with a few important changes. For one, I quickly realized that the Elements class needs to have access to the ID array and Node_Array. I therefore added static members to do just this, as well as a function to set the up.

By default, Node_Array and ID are set to NULL (see the top of Elements.cc). I can then use the value of these arrays to determine if Elements can be set up or not (see the explanation at Elements.cc:100).

I explicitly disabled the = operator. I added a deep copy constructor (not sure if it'll ever be used but whatever).

I still need to add members to set Ke, among other things. A printing function would also be useful. I also think that a function to access Local_Eq_Num_Tol_Global_Eq_num would be useful.

I also want to add an Error file. I think that it's time for me to have actual error handling. I want to do this using a gigantic Enumerator that stores all possible error states. Void functions will then return an error state. If zero, the program will go on. If not, then the error state can be sent to an "Error_Handling_Function" to determine what went wrong (and hault the simulation)!

This code compiles, but has not been tested (I need to write an element test function).