Fixed Node error
There was a big problem with the way that the Node class had been set up. Basically, the current position was never set unless a Node had BC's supplied to it. In my test, however, I only gave BC's to nodes with fixed BC's. This made me realize that my implementation was confusing. To fix it, I rewrote the node class.

I got rid of several functions, including Set_BCs and Set_Current_Position (private) along with a few members (BCs have been set + BCs). I replaced Has_BC's with a new member, Fied_Pos. Now, if you want to set a Node's position, you must specify the Original position and the BC's (just a booolean array. ith component is true if ith component of position is fixed).  The origional position should include the BC.

Once run, you will not be able to modify any of the components of the position that have Fixed_Pos[i]== true.

I basically realized that there was no need to have the BC array... I just needed an original position + information on which BC's can't be modified.

This new code is, in my opinion, much easier to understand.

I had to update the Errors (see Errors.h, Errors.cc) to work with this new implementation + Modify some stuff in the Elements class. All said and done.

With these changes made, the code runs correctly!

I plan on making the BC's argument for "Set_Original_Position" a defulat argument (default to all false). However, I don't think that I can do that with a class... it may be time to get rid of the Array_3 class (or just change that argument to a set of 3 booleans... who knows what I'll do). Expect changes in the next commit.

This code compiles and runs!