Added Element Errors
I added error handleing for the Error class. See Errors.cc and Errors.h for detials.

I choose to change the element class so that the "Node_ID" method returns the ID through a second argument, rather than through the return variable. I made this change because I wanted this method to be able to return an Error if the requested index is out of bounds.

I plan on rewirting the node class getter members so that they also return Errors. This will require rewriting a few members.

Because this new notation is a bit more confusing (it's not clear which arguments are inputs, which are outputs, and whatnot), I plan on redoing my comments (in the header files) to make it explicitly clear which arguments do what (in, out, or in/out). Expect that change (along with the Node member rewrites) in the next commit.