Updated comments + Rewrote some node functions + Added Func ptr to Element class
I added an "Intent comment" to each argument of each method in the Node and Element header files. These intent comments specify what the intent of the associated arugment is (either read, write, or read/write). I did this b/c basically every function now returns errors, so outputs go through the arguments (not return variables).

To streamline this whole process, I rewrote a few of the Node functions. Namely the getters. These methods now return the position through a reference argument. The functions now return an Error (Success if used properly, but NODE_NOT_SET_UP if the node hasn't been set up).

I also renamed the awful "Origional_Position_Has_Been_Set" member in the Node class to "Node_Set_Up"

Finally, I added a new static member to the Element class, the "F" pointer. This is a pointer to the function that will actually perfom the integration (that us used to assembel Ke). The idea is that it accpets a node ID and component number for the two things that you want to integrate (therefore 4 unsigned arguments) and returns the associated integral. See my FEM codes for detials.

This code runs. However, I need to write new Tests to make sure that the Node getters work as intended. I also need to finish writing the Element class (write a function to assemble Ke  + write a function to map Ke to K). Actually... now that I think about it, the Element class probably needs another static member, the K member.... Once I have all that written, I really need to write and perform some tests on the Element class.

This code compiles but has not been tested.