Added K static member to Element class
I added another static member to the Element class. This static member points to K, the global stiffness matrix. I realized that there needed to be a pointer like this for each element to be able to populate K.

I plan on adding two functions to the Element class. One that will populate Ke and another that will move Ke into K.