Wrote a "Populate Ke" function
I wrote a function to Populate Ke. this is basically just an implementation of the outline on the FEM document.

I made a few changes to the Element class beyond just adding this method. For one, I realized that I needed another Element error, the "Element_Not_Set_Up" Error. I dediced to rename the Nodes_Set variable to Elements_Set_Up. This new name mirrors what I did for the node class + is more in line with what the variable is actually used for. The new error is triggered whenever the user tries to access some node info (ID, or set Ke) before the Element has been set up (by the Set_Nodes method).

I plan on adding a "Move Ke into K" function + to add explicit "Assumption" comments to the start of some other methods in the Element class (see the new Populate_Ke function for an example of what I mean).

This code compiles but has not been tested.