Added assumptions to Node + Element methods
I added proper assumption comments to the Node and Element class methods (all of them that have assumptions). An assumption is just something that must be true for the method to work properly. For example, the current position method (node class) assumes that the Node has been set up.

I also added the Matrix template class to the Forward_Declarations file. I'm not currently using this file, but it may become important at some point, so It's probably a good idea to keep it updated.

Next, I plan on rewriting the Element class in terms of Matrix objects.

This code compiles and runs