Fixed several bugs, got Element test running
I finished the first version of the Element test. This revealed several  bugs.

To finish the Element test, I added code that makes an array of elements, sets up each element's neighbors, populates each element's stiffness matrix, and then moves those stiffness matricies into K. Interestingly, I was able to do all of this in a single for loop. This is because the Elements can be populated sequentially. This should, in theory, allow for more parallelization (though the process of moving Ke to K must happen sequentially to avoid a race condition). I then added code to print out K (just to verrify that everything working)

Once this was done, a whole slew of bugs presented themselves. First, I realized that using 8 elements (3x3x3 array of nodes) was wayyyyy too much to print out. Thus, I bumped Nx Ny and Nz down to 2 (from 3) so that I only have 1 element. I also got rid of the BC's (just for testing purporses). This made K a 24x24 matrix.

This made K a reasonable size, but revealed all kinds of problems. First, most of the elements in K were empty, which seems very odd (the way that my "integral" function is set up should fill every element of K to 1). This implies that something was wrong with moving Ke to K. The issue was that I had swithced the way that ID was set up. I now accessed it by Node number (for the row) and then component number (column). I had to make this fix (see Element.cc:127).

Once that issue was fixed, K was lookig better, but still pretty bad. The entire first column of K was zero. The issue, once again, had to do with ID. I stupidly set up the global equation numbers to be 1 indexed (first equation number is 1), so the first element of ID was 1 and not 0. This obviously created a problem, since all arrays in C are 0 indexed. the ID array was mapping into the wrong elements of K. To be entirely honest, I'm not sure why this didn't cause a segmentation fault (it should have). To fix it, I had to make some pretty serious changes. with a 1 index, if a component is fixed, we can fill that component's associated cell in the ID with 0. This means that every element of ID will be an unsigned integer. However, if we are zero indexed, then we need to use something like -1 to denote an fixed component. This forces ID to be an integer matrix, not an unsigned matrix. This required several changes (See element.h, Set_element_Static_Members, and more).

Making this change fixed the "column of zeros" issue in K. Every element of K now has something in it. However, the diagional elements of K had a 3 in them, not a 1. Since F always returns 1, this implied that I was populating some components of K multiple times. Something was wrong with how I moved Ke to K. The issue was on line 222 of Element.cc. When moving the off diagional cells of Ke to K, I started at Col = Row, not Col = Row + 1. Therefore, my "off diagional loop" was actually going through all of Ke. This triple counted the diagional entries. I fixed this. I also realized, however, that my code here was going in Row major order, even though Ke is a column major matrix. I fixed this by reversing the order of the loops (hopefully improving runtime!)

With these changes, K became a matrix of all 1's, just as I expected.

After fixing these bugs, I changed the way that Elements are set up to improve runtime. To do this, I first get the global eq number. I then check if that is -1 to determine if I need to set Local_Eq_Num_To_Global_Eq_Num and such. See Element.cc:130 for the changes. The old system involved checking the current componet of the current node was fixed, then fetching the component from ID. This meant that I was never actually using the -1 condition in ID (if component is fixed then ID component is -1) + was doing two operations when I could do just 1. This change is likely more robust (since it relies entirely on the ID matrix) + should be a little quicker.

From here, I want to make the tests better. For one, my test right now only has 1 element. This is because any more is too much for the command line. Thus, I probably want to print K to a file and then increase the number of elements. This will allow me to see if K is indeed sparse (as it should be). I also need to test how K works with BC's. Further, I want to split the Tests.cc file into serveral files (one for each type of test) and then move those to a separate file. I also want to look into turning parts of the Element test into functions (which will eventually be used in my simulation code).

This code runs and has been tested.