Ke is now always 24x24
I realized that there was a flaw in my understanding of the finite element method. I thought that the local stiffness matrix would, in general, have ≤ 24 rows/cols (since any local equations corresponding to fixed global equations would not show up in Ke). I belive that this understanding was wrong.

Instead, Ke should always have 24 elements, all of which will get popuated (this seems inefficient, but oh well). The element keeps track of which local equations are fixed. When Ke is moved to K, any components corresponding to fixed equations are skipped/not moved to K (since they have nowhere to go).

This realization forced me to change the mechanics of the Element class. I no longer needed a Num_Local_Eq, since that was always 24. I also no longer needed to dynamically allocate Ke, Local_Eq_Num_To_Node or Local_Eq_Num_To_Global_Eq_Num since these array's sizes are also already known.

I did, however, need some way of determining if a component was fixed or not. I did this using a new static memeber, FIXED_COMPONENT. This is initialized to -1 even though FIXED_COMPONENT is an unsigned integer. This is done because (unsigned)(-1) corresponds to a massive integer, larger than any possible (realisitally) global equation number.

I realized that the same trick could be played with the ID array, so so I switrched ID back to a matrix of unsigned integers/played the same -1 trick (see Element_Tests.cc:46, Element_Tests.cc:84).

With these changes, I had to redo my destructor (since there are no longer any dynamically allocated arrays to free) and many of the Element methods. See Element.cc for the detials. To make a long story short, if a local equation corresponds to a fixed component then the corresponding elemnet in Local_Eq_Num_To_Global_Eq_Num gets set to FIXED_COMPONENT. We can use this constant to check if a particular component of Ke should be moved to K or not. This is explained much better in the "Robert's FEM planning" document on google drive.

This code has been tested a few times and it appears to work.