Removed Local_Eq_Num_To_Node, added Xa, Ya, and Za
I realize that I don't actually need the Local_Eq_Num_To_Node member. I did, however, need some way of determining the position of each node.

Thus, I removed the Local_Eq_Num_To_Node Member and added 3  new members, one for each coordinate of the node positions. The ith component of Xa is the x coordinate (spatial) of the ith node in the Node list. Ya and Za are defined similatly, see Element.cc:102

To be able to populate Xa, Ya, and Za, the element class needed access to the Node array. Thus, I reintroduced the Nodes pointer for the Element class (this gets passed to Set_Element_Static_Members). See Element.cc:333 for detials.

I also realized that there was a serious bug with my code. I was declaring Na, Na_Xi, and such as local global variables, not as a part of the element class. I fixed this. See Element.cc:16-17 and Element.cc:384-396.

I believe that this code is working (I need to test it more to be sure, thus the Element_Monitor is defined).

In my next commit, I plan on adding code to calculate the coefficient matrix, moving me closer to actually being able to calculate Ke.