Added Calculate_Coefficient_Matrix method (private)
I added a method to calculate the Coefficient matrix and Jacobian determinant for a given integration point. This was done using the proceedure outlined on pages 149-150 of Hugues' book.

I decided to make this function a private function. This is because this method really just helps calculate Ke. There is no reason to calculate the Coefficient matrix outside of Calculate_Ke. Therefore, I thought that it made sense for this method to be private. Ths implementation of this code can be found att Element.cc:185-262.

I added an outline for this function to my FEM planning document.

I also added a new Element error, the "Integration_Point_Index_Out_Of_bounds" error. This is done to handle the case when the Integration point index is >= 8. See Errors.h and Errors.cc for the implementation. This error can be thrown by the Calculate_coefficient_matrix function (see assumption 1, Element.cc:190).

I also reorganized the Enumerators and Switch statements for the Error handling code. This is all designed to  make the code more readable.

I also modified the error handling system in the Element class. If an error is encountered (an assumption fails) then the function prints the name of the function that triggered the error before returning the error code. This should, in theory, make it easier to trace the origin of an error when one is encounterd.

This code compiles, but has not been tested. In my next commit, I want to try calculating the Coeffieint matrix for an element or two to verrify that it at least seems to be working properly. I also want to continue deveoping the Populate_Ke code. Next, I want to add code that calculates the partials in the spatial domain (Na_x, Na_y, Na_z). Later, I want to write code to calculate each Ba/map them into B (this will again be done in a private function). I also want to look into replacing Static pointers with references.

This code compiles but has not been tested.