Got rid of Matrix's set up method
I really have no idea why this method still existed... I probably thought that it was really clever at one point, then never really used it the way that it was intended ot be used. Whatever the reason was, this method was dumb. It was only being used twice in the entire program. In both instances, using Set_Up wasn't necessairry. See Matrix_Tests.cc:13, 41.

As such, this method no longer had a reason to exist, so I got rid of it. This allowed me to make Memory_Layout a constant (see Matrix.h), which sould improve runtime in a few spots.