Started rewrite of matrix class + error handling system
Ke is a mess. There is no denying it. The only solution, in my mind, is to make a matrix class that can do matrix-matrix multiplication. This requires seriously overhauling the existing matrix class. In particular, I want to add a * operator, a = operator, a move constructor, a move = operator.

The * operator needs to be smart in the sense that it should be able to use the fastest method for calculating a matrix product given their memory layouts (row or column major). The same goes for the other operators.

I also want to redo the entire error handling system in my program. I want it to be done in terms of exceptions, not error codes (this will let me write some far more readible code).

All of this will take time....

On the bright side, I did manage to start. I added exception handling for the matrix class (See Errors.h, Errors.cc, and Matrix.cc). I also began writing a * operator, but stopped b/c I ran out of time.

I also added a new "Matrix_Errors_Test" to verrify that the exceptions are working. See tests.h for details.

Expect more changes soon.