Matrix-Matrix Multiplication + Matrix Move Constructor
I added a Move-constructor for the Matrix class. This is pretty standard, it just hands off the Matricie's Dynamic array. See Matrix.cc:25-36

I didn't need a Move assignement operator. Thus, I explicitly deleted this method. See Matrix.h:40.

I also wrote Matrix-Matrix Multiplication. The specific algorithm that is used depends on the memory layout of both matricies involved. Doing it this way allows me to optimize memory usage, regardless of the memory layout. See Matrix.cc:93-238 for details.

A new method called Zero was added to help compute the matrix-matrix product. This method, as the name suggests, zeros out the elements in a matrix. I decided to make it a part of the public interfact b/c why not?

I wrote a new test to make sure that the new methods worked actually work... which they appear to.

I also found and corrected a small typo in Ke.cc