Rewrote matrix exception (+added move = operator)
I added a move assignment operator. I did this so that I could do things like line 250 of Ke.cc. Here, JD has already been initialized, the code here reassigns or "updates" JD. The old data in JD is not needed and J*D (the thing we're setting JD equal to) is an R-value. As such, this is a great canidiate for the move assignment operator.

The move assignment operator definition is pretty standard, though it has a few quirks. For one, Num_Cols, Num_Rows, and Memory_Layout are fixed in the Matrix class. This is done to make the code more explicit/easier to debug. Once a matrix is defined, it's dimensions don't change. the dimensions/layout prescribed at initialization presist forever. As such, the move assignment operator can not change Num_Col/Num_Row/Memory_Layout. Thus, this method checks that Other is compatible with *this, which basically means that it has the same dimensions/layout as *this. If this is not satisified, then a "Matrix_Dimension_Mismatch" exception is thrown.

That brings me into the other major change of this commit, matrix exceptions. The matrix exceptions that I had worked, but they were kinda annoying to work with. In particular, each exception class had a different interface... keeping track of all of that was annoying. I also realized something: When my code rasised an exception, I know exactly why that exception was raised. In other words, explaining the exception should take place when the exception is thrown, not when it is processed. Thus, I overhauled matrix exceptions... There is now a base Matrix_Exception class. This class has a constructor, a what() method, and a string member. Matrix exceptions are initialized by supplying the string (the error message). The what() method returns the error message as a char array. Each specific exception is derived from the base Matrix_Exception class. Using inheritence actually allows me to do a lot of cool things. For one, each drived object is a Matrix_Exception. Thus, I can write a "catch-all" handler to catch any kind of Matrix_Exception. All specific matrix exceptions (like dimension mismatch and such) will be caught by this handler. At the same time, however, I can write handlers for specific derived class exceptions. This allows for flexible error handling code!  See Errors.h for implementation of the new matrix exception classes. I plan on using this system to define the exception classes for Node and Element.

With the new matrix exceptions, I had to rewrite the "throw" parts of the Matrix class. This involved mult-line strings, char buffers, and all kinds of fun! See Matrix.cc for detials.

This code runs and handles exceptions as expected.

Next, I want to overhal error handling for the Node and Element classes (along with tests to validate the changes). After that, I will rewrite Ke to be easier to read (using the new Matrix class methods!). Once I have done that, I will begin adding new functionality to the code (such as a solver for Ke*d = F and functions to input/output data).