I added Node and Array ecxceptions (but did not implement them)
I wrote exception classes for Nodes/Arrays. I have not, however, implemented these classes. Expect implementation in the next commit.

Note: I decided to not make an exception class for the Node INDEX_OUT_OF_BOUNDS error code. The reason is that this error code gets thrown whenever the user tries to access an invalid component of an array... this is really an error with the Array class, so it should be treated as an Array exception, not a node one. Thus, I decided to make a set of Array_Exceptions (but have some node methods throw them).

Expect all of this to be implemented in the next commit.