Changed Ke code to use Matrix methods!
I cleaned up Ke. For one, I modified it so that it uses the matrix operators rather than nasty (and easy to mess up) for loops.  See Ke.cc for detials.

I reran element (using the new code) and, according to diff, the output is the same :D. :+1: :+1: :+1:

Unexpectedly, the biggest change in this commit was the addition of a "Print_Matrix_Of_Doubles" method. This function, as the name suggests, is used to print out a matrix of doubles. I wrote this to clean up the monitor parts of the element class. Previously, the same code (to print out a matrix) was repeated like 6 times... once for each matrix... that was ugly. I fixed that. The method prints in exponential format by default and has configurable width/precision parameters. See Core.cc.

The biggest issue that I had with with Print_Matrix_Of_Doubles was deciding where to actually put it. Placing it in the matrix file caused problems (since that file defines a template and is therefore included in several places). As such, I sorta awkwardly declared it in Elment.h and defined it in Core.cc.... I should probably find a better home for this function.. idek... on the one hand, the element classes are the only methods that call it, so it makes sense to place this method in the Element class folder, even if it does seem out of place.

Next, I want to add exceptions to the private members of the element class, then move onto solving Ke*d = F (this will probably be done using a library).