Added Private exceptions to private element methods
I added exception handling to Calculate_Coefficient_Matrix and Add_Ba_to_B. See Ke.cc for details. It should be noted that I couldn't find a way to test the validity of assumption 2 in Add_Ba_To_B (see Ke.cc:147), so I left this one as is.

I also tweaked the formatting of the Monitor parts of the Element class (which is why the other files changed). This code runs!

Next, I want to add a += method to the Matrix class (to clean up Ke.cc a little bit more). Once that is done, I will find a way to solve Ke*d = F.