Added += operator to matrix class
I added a += operator to the Matrix class. As usual, the way that this is done depends on the memory layout of both matricies involved. I also added a test to Matrix_Tests.cc to test the new method. It appears to work.

With tests validated, I converted Ke.cc to use the new method. This should improve performance and make that line exception safe.

With that, I think that my rewrites are pretty much done. Ke is now much easier to read, exception safe, and easier to maintain (since most of the work happens in the Matrix class now, meaning that it should be a lot easier to isolate any potential issues).

I will now look for a way to solve Ke*d = F nd then move onto printint out the results (as well as implementing a multi step process and force calculator).