Merged Matrix.cc and Matrix.h
Matrix is a template class. I tried placing the class declaration in Matrix.h and the definition in Matrix.cc, as if the Matrix were just a regular class... That approach worked, but it was confusing (Matrix.h included Matrix.cc at the end, which is just weird). Thus, to clear up confuson, I merged the two files.