Added Pardiso Tests
I added code to test Compress_K. Since I expect there to be more Pardiso tests in the future, I wrote a whole new file (Pardiso_Tests.cc) to see that everything works.

So far, things appear to be working (woo-hoo).

In the next commit, I will be adding a bin directory and modifying the makefile so that it dumps object files (and the exectuable) into that directory rather than the main directory.