updated makefile
I changed a bunch of things with the makefile. First, and most importantly, the old makefile would always run each rule. The reason is that the rules needed to be prefixed with ./obj/. I fixed that.

I also changed the compiler from g++ (which is really just an alias for clang) to g++-9 (which is the ++ that I got from homebrew) so that I can use openmp.

I also renamed _OBJS to OBJS and OBJS to PATH_OBJS, I think these variable names better reflect what these things are.

Finally, I changed the all rule so that it simply depends on each object file and the target (the FEM exectuable). I wrote a new rule to make the target. This way, all contains no commands and therefore does nothing if everything is compiled (which is what it should do).