Added a bin directory + Run script
I added a bin directory. The FEM object file is now compiled to that folder, rather than in the main directory.

I also added a short run script. This sets up the relevant variables and runs FEM. This is a modified version of the run script that I wrote for my Pardiso test repo.

I made corresponding changes to my makefile.

This code runs and compiles, though I have yet to test the Pardiso_solve method. I plan on doing that in the next commit (which will let me move onto verrifying that my code is working).