added VPATH + automatic variables to makefile
I finally learned how to use automatic variables and VPATH in makefiles. I updated the Makefile to use both of these things.

The makefile does the same thing, it's just much easier to read now.

I still want to figure out some way to get rid of the obj/ on each rule, but honestly that's not a big deal. The important part is that the makefile is much smarter now.