Updated Compress_K to work with both memory layouts
The old Compress_K code was set up for row major matricies. The issue is that K (and others) are column major. I modified the code so that it is optimized to work with both row-major and column major matrices. See Compress_K.cc

I also got rid of the "Update_Position_Using_BCs" method of the Node class (which has long been slated to be eliminated).

The rest of the changes in this commit are minor syntax/style changes.