Redid Node class (and added some stuff to Element Tests)
I basically totally redid the Node class. The Node class was helllla out of date. For one, it still had two positions, Original and Current. These are completely outdate. There's just one Position.

I also changed the "Update_Position" method to "Update_Position_Component". The new name better reflects what the function does.

I also got rid of BC. These aren't needed. The BC's are used to prescribe the Position, they are not their own thing.

I also changed how BC's can be set. I changed it so that BC's can be updated, but that a position can not be updated if it is a BC. SEe Node.cc for changes.

I also got rid of the Get_Current_Position/Get_Original_Position methods in favor of a new Get_Position_Componnet method. Nothing was using the old methods, they were kinda useless. The element code does, however, take specific positions from the Node, so I added this method to allow the element class to do that.

This allowed me to get rid of the friend relationship between the Node/Element class (The Element class needed access to the private members of Node in the past b/c the Node public methods were so useless. This should help improve code integrety/debugging.

Making all of these changes forced me to rewrite a lot of the Node_Tests.

I also modified the Elemnet_Tests to include a solving step. Doing this forced me to incldue Pardiso_Solve in Element_Tests.h, which revealed an issue: Pardiso.h didn't have a compile guard and it contained source code. I fixed this by making a new file, Pardiso_Error.cc, to house the Error handling method (Report_Pardiso_Error) and then added a compile guard to Pardiso.h. This got rid of the linking error.

I also disabled several monitors and the pardiso msglvl variable so that the output of the Element Tests is managable.

This code compiles and runs.

In my next commit, I want to change the Element_Test so that it prints the new node positions (this should be an easy change). Then I will move into verrification mode.