Redid Node class (Again....)
I realized that I was doing dumb things in the Node class. The BC's are DISPLACEMENT BC's, not Position BC's. I also realized that the prescribed positions in the F calculation are the DISPLACEMENTS not the positions. With these revelations, I realized that I needed to add a displacement component to the Node class.

Amazingly, the existing Element tests only set the presceibed BC's on nodes that had position components of zero (whcih si the same as their displacements), so I didnt notice any issues.

To fix this issue, I added a displacememt member to the node class. The set BC methods set the displacements (and lock that component off from being set directly). I added new methods to directly set and access components of the displacement. See Node.h and Node.cc for all the detials.

In the next commit, I am going to change the element code so that it uses the displacement, not the positions, when setting the presceibed positions.