presceibed_Displacements + updated Element tests
I replaced presceibed_positions with presceibed_displacements. The new variable is set based on the displacement member of the node class, not the position (as it should be).

See Core.cc:149 for the detials.

I also modified the printout section of that file.

I then made some big updates to the element test file. First off, the position should ALWAYS be set (since the BC members set the displacements, not the positions). I also added a new condition to stretch certian parts of the cube. See Element_Tests.cc:292

I also updated how the nodes print out after the simulation is done.

With all of these changes, the code runs and seems to be giving correct numbers (woo-hoo). When I stretch one face of the the cube, the rest of the cube sort of pulls inward (as it should).