Added in code to read in boundaries
I modified the inp_reader so that it can now read in boundary data. To do this, I needed to create a new inp_boundary_data list (see source/IO/inp_reader.h). I also modified a few bits of the inp_reader (like throw-away read-in fields, see inp_reder.cc::63 and such)

In the next commit, I plan on adding a simulation folder (and basing it off of the code in the Element_Test file).

I also need to write a method for the element class that updates the node BC's (which will be needed for calculating the force vector after each step. This will change Fe, but not Ke. Thus, after each update, F will need to be reclalculated, but not K)