fixed bugs in vtk_Writer.cc
I modified the Element tests so that it now outputs to a vtk file rather than to the consul.

This gave me a way to test that the new writer was actually working.

Turns out it wasn't... the file path was wrong and I forgot a newline chracter at the end of the header. I fixed both of these issues. See vtk_Writer.cc.

With this, I got the output working! the vtk writer appears to give correct output (though I have yet to verrify it)!