Added wedge element test
I added a test for wedge elements! They seem to work!

The test basically creates a triangular prism and then pulls the top (the crest of the prism). This causes the prism to deform/elongate.

I added this test to Element_Tests.cc. I decided to renamed the old "Element" method to "Brick_Element" (since that's what it's testing).

Working out which nodes are associated with which element was very tricky. I didn't try to explain in code (since it doesn't really make sense without paper/pictures.... oh well). Either way, it seems to work.

I also added in a "TEST_MONITOR" to toggle printing out stuff during tests.

With all of these changes, my code seems to run (I checked it visually in paraview and everything!).

In my next commit, I plan on adding support for reading/writing of wedge elements.