Wrote a "Contains" method for reading (+ added tests)
I plann on adding IO support for wedge elements. Doing that requires me to be able to identify certian substrings within the input buffer. I've done this before, but the location of the substrings has always been very predictable. That may not be the case going forward. As such, I wanted a foolproof method for identifying substrings in a string. To do this, I wrote the contains method (See inp_reader.cc).

This method is pretty straight forward. It accepts two null terminated strings, a buffer and a word, and tries to identify the word in the buffer. To do this, it loops through the characters in the buffer until it finds one that matches the first character of the word (or reaches the end of the buffer). It then checks if successive letters in the word/buffer match (or either ends). If every letter of the word can be found in the buffer, then it returns true. Otherwise it returns false.

I also wrote some basic unit tests to make sure that this function works (see IO_Tests.cc).

Based on these tests, the new function seems to work.

In the next commit, I plan on rewriting the inp reader in terms of this function (rather than those stupid "check 7 letters in a row" things). I will then add support for reading wedge elements.