Added a "Start_At" parameter to Contains
I realized that, when checking if a word is contained in a buffer, I often know that the word won't be contained in the first n letters of the buffer (for some natural number n). To improve efficiency, I added an optional "Start_At" parameter to the Contains method. Basically, if a non-zero value is passed for Start_At, then the Contains method will search for word starting at index Start_At of Contains (it will not look for a match in the first Start_At-1 characters of Buffer, even if a match does exist in those characters).

I wrote some extra tests to see if the function works with this new paramater, it appears to do so!